server/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.20.0
python-jose==3.3.0
redis==4.3.5
requests==2.31.0
prometheus_client==0.16.0